New method for unit start that reports whether the launch was successful
Adds a `startAndWait` method in the TypeScript `Unit` class that starts a given unit/service and waits a certain amount of seconds between active state changes to determine and report whether the service start was successful or not.

This is implemented in the Rust library as a `unit_start_and_wait` method of the `System` struct.

Change-type: minor
Signed-off-by: Zahari Petkov <zahari@balena.io>